Fix opposite edge brick disappearing bug
Since we are doing rough collision detection, which is not perfect, when
ball goes off the edge, the brick from the previous row on the opposite
side disappears.

This fixes it by refactoring the control for the ball to be inside the
brick grid.

Also name the controls as `gotBrickLeft` and `ballInsideBrickGrid` for
better legibility.