Fix covered edge corner case
This fix refers to Section 5 Lecture 44.

When there's a lonely brick and the ball hits its corner on a 45 degree,
ball should bounce back to the same direction it came, that means
flipping both x and y velocity.

But if the brick has a neighbor, then it should bounce like when
it touches the middle of the brick.